Conflicts:
	src/main/java/persistance/dao/impl/UserDaoImpl.java
	src/main/java/service/TestService.java

+++ src/main/java/persistance/dao/GenericDao.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface GenericDao <T> {
     public void save(T entity);
-    public void removeById(int id);
+    public void removeById(long id);
     public void update (T entity);
     public T findById(long id);
     public List<T> findAll();
